MIGENG-151 moving WorkloadInventoryReportModel to the right package
